add route to one product

+++ server-side/app.py
@@ -40,10 +40,36 @@ def add_product():
     products.insert_one(item)
     return json.dumps(item, default=str)
 
-@app.route('/produtos/<string:id>', methods=['DELETE'])
-def delete_product(id):
-    products.delete_one({'_id': ObjectId(id)})
-    return "<h1>Teste</h1>"
+@app.route('/produtos/<string:id>', methods=['DELETE', 'PUT', 'GET'])
+def delete_or_update_product(id):
+    if request.method == 'DELETE':
+      products.delete_one({'_id': ObjectId(id)})
+      message = "Produto deletado"
+      return json.dumps(message)
+
+    if request.method == 'PUT':
+      data = request.get_json()
+      nome = data['nome']
+      quantidade = data['quantidade']
+      preco = data['preco']
+      codigo = data['codigo']
+      categoria = data['categoria']
+
+      newvalues = { "$set": { "_nome": nome, "_quantidade": quantidade, "_preco" : preco, "_codigo" : codigo, "_categoria": categoria } }
+      products.update_one({'_id': ObjectId(id) }, newvalues)
+      message = 'Produto atualizado'
+      return json.dumps(message)
+
+    if request.method == 'GET':
+      item = products.find_one({'_id': ObjectId(id)})
+      return json.dumps(item, default=str)
 
 
-app.run(debug=True)
+
+@app.route('/encontrar/<string:id>', methods=['GET'])
+def find_product(id):
+    item = products.find_one({'_id': ObjectId(id)}) 
+    return json.dumps(item, default=str)
+
+
+app.run(host='localhost', port=int('5000'), debug=True)
